requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.3.0
celery==5.3.6
redis==5.0.4
django-celery-beat==2.6.0
sendgrid==6.11.0
gunicorn==22.0.0
whitenoise==6.7.0
django-cors-headers==4.3.1
